minor code adjustments and add print log

+++ code/copy_setups_to_trial.py
@@ -14,8 +14,14 @@ def run():
     for key, value in paths.__dict__.items():
         if key.startswith('GENERIC'):
             dest_path = os.path.join(paths.TRIAL_DIR, os.path.basename(value))
-            shutil.copy(value, dest_path)
-            print(f"Copied {key} to {dest_path}")
+            if not os.path.exists(dest_path):
+                shutil.copy(value, dest_path)
+                print(f"Copied {key} to {dest_path}")
+            else:
+                print(f"{dest_path} already exists, skipping copy for {key}")
     
     # Copy CEINMS files to the trial directory
     
+
+if __name__ == "__main__":
+    run()
